+++ common/simulator/spring.h
@@ -6,8 +6,8 @@
 #include "math.h"
 
 struct Spring {
-	uint m0;
-	uint m1;
+	uint16_t m0;
+	uint16_t m1;
 	float rest_length;
 	float mean_length;
 	Material material;
